backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
qdrant-client==1.7.0
google-generativeai>=0.8.3
openai==1.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytesseract==0.3.10
pdf2image==1.17.0
PyPDF2==3.0.1
tenacity
pypdf
requests
python-dotenv==8.2.3
